Project specific vim settings can be sourced.
A file '.vimrc_local' when present in the current working directory will
be searched for and sourced when found. In this file, it is possible to
list settings that apply only to files in that directory, possibly
overriding defaults. A usecase is to fulfil the need for modeline-like
behavior (without actually using modelines, whose use has been
historically discouraged). For example: specifying a project specific
makeprg command for .c files in that directory.

Another use case is to locally enable modelines (to set file specific
behavior) only for files in that directory with the assumption that the
directory is from a trusted source (which is certainly the case if you
yourself are the author).